consolidates components, cleans up Root.jsx
