+++ src/styled/StyledLink.jsx
@@ -1,0 +1,11 @@
+import styled from '@emotion/styled';
+import { Link } from 'react-router-dom';
+
+const StyledLink = styled(Link)`
+    text-decoration: none;
+    color: rgb(255, 255, 255);
+    font-weight: bold;
+    /* color: #64dfdf; */
+`;
+
+export default StyledLink;
